+++ lib/zendesk_api/collection.rb
@@ -1,14 +1,14 @@
 require 'zendesk_api/resource'
 require 'zendesk_api/resources'
 require 'zendesk_api/search'
+require 'zendesk_api/pagination'
 
 module ZendeskAPI
   # Represents a collection of resources. Lazily loaded, resources aren't
   # actually fetched until explicitly needed (e.g. #each, {#fetch}).
   class Collection
-    DEFAULT_PAGE_SIZE = 100
-
     include ZendeskAPI::Sideloading
+    include Pagination
 
     # Options passed in that are automatically converted from an array to a comma-separated list.
     SPECIALLY_JOINED_PARAMS = [:ids, :only]
@@ -114,30 +114,6 @@ module ZendeskAPI
     def count!
       fetch!
       @count || -1
-    end
-
-    # Changes the per_page option. Returns self, so it can be chained. No execution.
-    # @return [Collection] self
-    def per_page(count)
-      clear_cache if count
-      @options["per_page"] = count
-      self
-    end
-
-    # Changes the page option. Returns self, so it can be chained. No execution.
-    # @return [Collection] self
-    def page(number)
-      clear_cache if number
-      @options["page"] = number
-      self
-    end
-
-    def first_page?
-      !@prev_page
-    end
-
-    def last_page?
-      !@next_page || @next_page == @query
     end
 
     # Saves all newly created resources stored in this collection.
@@ -323,11 +299,6 @@ module ZendeskAPI
       map(&:to_param)
     end
 
-    def more_results?(response)
-      Helpers.present?(response["meta"]) && response["meta"]["has_more"]
-    end
-    alias_method :has_more_results?, :more_results? # For backward compatibility with 1.33.0 and 1.34.0
-
     def get_next_page_data(original_response_body)
       link = original_response_body["links"]["next"]
       result_key = @resource_class.model_key || "results"
@@ -344,35 +315,6 @@ module ZendeskAPI
 
     private
 
-    def cbp_response?(body)
-      !!(body["meta"] && body["links"])
-    end
-
-    def set_cbp_options
-      @options_per_page_was = @options.delete("per_page")
-      # Default to CBP by using the page param as a map
-      @options.page = { size: (@options_per_page_was || DEFAULT_PAGE_SIZE) }
-    end
-
-    # CBP requests look like: `/resources?page[size]=100`
-    # OBP requests look like: `/resources?page=2`
-    def cbp_request?
-      @options["page"].is_a?(Hash)
-    end
-
-    def intentional_obp_request?
-      Helpers.present?(@options["page"]) && !cbp_request?
-    end
-
-    def supports_cbp?
-      @resource_class.const_defined?(:CBP_ACTIONS) && @resource_class.const_get(:CBP_ACTIONS).any? { |supported_path| path.end_with?(supported_path) }
-    end
-
-    def first_cbp_request?
-      # @next_page will be nil when making the first cbp request
-      @next_page.nil?
-    end
-
     def get_resources(path_query_link)
       if intentional_obp_request?
         warn "Offset Based Pagination will be deprecated soon"
@@ -387,27 +329,6 @@ module ZendeskAPI
         handle_search_export_response(@response.body)
       else
         handle_response(@response.body)
-      end
-    end
-
-    def set_page_and_count(body)
-      @count = (body["count"] || @resources.size).to_i
-      @next_page, @prev_page = page_links(body)
-
-      if cbp_response?(body)
-        set_cbp_response_options(body)
-      elsif @next_page =~ /page=(\d+)/
-        @options["page"] = $1.to_i - 1
-      elsif @prev_page =~ /page=(\d+)/
-        @options["page"] = $1.to_i + 1
-      end
-    end
-
-    def page_links(body)
-      if body["meta"] && body["links"]
-        [body["links"]["next"], body["links"]["prev"]]
-      else
-        [body["next_page"], body["previous_page"]]
       end
     end
 
@@ -567,17 +488,6 @@ module ZendeskAPI
     def assert_results(results, body)
       return if results
       raise ZendeskAPI::Error::ClientError, "Expected #{@resource_class.model_key} or 'results' in response keys: #{body.keys.inspect}"
-    end
-
-    def set_cbp_response_options(body)
-      @options.page = {} unless cbp_request?
-      # the line above means an intentional CBP request where page[size] is passed on the query
-      # this is to cater for CBP responses where we don't specify page[size] but the endpoint
-      # responds CBP by default. i.e  `client.trigger_categories.fetch`
-      @options.page.merge!(
-        before: body["meta"]["before_cursor"],
-        after: body["meta"]["after_cursor"]
-                           )
     end
   end
 end